authorization system updates (#2868) (#2935)
* first commit



* debug



* authz api first try



* service api done



* change authz check order to avoid panic



* change authorization logic to avoid possible panic



* build module done & merge user client for picket(aslan)



* fix policy(aslan)-user dependency



* delivery module done



* moving code



* debug



* update collaboration mode apis



* env half done



* environment part fully done



* project done



* remove unused package



* debug logs & some system level authorization problems



* fix project list logic



* added query logic in mongodb



* fix collaboration mode creation error



* system authorization done



* collaboration mode fix for environment



* more collaboration mode fix for environments



* minor debug



* delivery center authz change



* workflow view authz fix



* workflow v4 filter & trigger filter



* debug logs for list workflows



* workflow module done for test



* fix init_info api authz problem



* fix production env creation problem



* fixed envcfg related api authz



* multiple authorization bugfix



* remove admin initialization from aslan, moving it to user



* remove useless code



* user db moved



* testing module done



* debug



* debug



* debug



* fix list test workflows



* added todo in comment



* system config related authorization done



* minor bugfix



* fix workflow plugin stuff



* user related api update



* workflow edit bug fix



* debug for list workflow task in product workflow



* added collaboration mode permission check in user and use it in aslan



* build service module authorization fix



* added list authorized env api to make life easier & multiple environment api authorization fix



* fix multiple permission problem with comments



* extend permission for multiple apis temporarily



* debug



* remove project-only authorization



* get user role update



* revert api deletion



* read-project-only permission change



* update collaboration mode handle logic



* bugfix



* debug info & warn log removal



* even more debug logs



* environment filter update



* fixed wrong list env logic



* fix panic bug



* added testing filtering



* debug logs & added project auth info for public project



* new permission logic for menu



* remove useless code



* remove bundle



* added some policies back



* remove useless code



* implement multiple authorization leaks on purpose



* remove authorization check for helm chart version



* debug



* update production env support for user



* multiple auth leaks



* debug



* production env permission



* ...'



* bugfix for permission



* debug update service



* debug



* remove useless comments



---------

+++ pkg/microservice/aslan/core/system/handler/basic_images.go
@@ -33,8 +33,21 @@ import (
 )
 
 func GetBasicImage(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
+	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
+
+	if err != nil {
+		ctx.Logger.Errorf("failed to generate authorization info for user: %s, error: %s", ctx.UserID, err)
+		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
+		ctx.UnAuthorized = true
+		return
+	}
+
+	// authorization checks
+	if !ctx.Resources.IsSystemAdmin {
+		ctx.UnAuthorized = true
+		return
+	}
 
 	ctx.Resp, ctx.Err = service.GetBasicImage(c.Param("id"), ctx.Logger)
 }
@@ -47,8 +60,15 @@ func ListBasicImages(c *gin.Context) {
 }
 
 func CreateBasicImage(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
+	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
+
+	if err != nil {
+		ctx.Logger.Errorf("failed to generate authorization info for user: %s, error: %s", ctx.UserID, err)
+		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
+		ctx.UnAuthorized = true
+		return
+	}
 
 	args := new(commonmodels.BasicImage)
 	data, err := c.GetRawData()
@@ -59,6 +79,12 @@ func CreateBasicImage(c *gin.Context) {
 		log.Errorf("CreateBasicImage json.Unmarshal err : %v", err)
 	}
 	internalhandler.InsertOperationLog(c, ctx.UserName, "", "新增", "基础镜像", fmt.Sprintf("label:%s", args.Label), string(data), ctx.Logger)
+
+	// authorization checks
+	if !ctx.Resources.IsSystemAdmin {
+		ctx.UnAuthorized = true
+		return
+	}
 
 	c.Request.Body = io.NopCloser(bytes.NewBuffer(data))
 
@@ -72,8 +98,15 @@ func CreateBasicImage(c *gin.Context) {
 }
 
 func UpdateBasicImage(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
+	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
+
+	if err != nil {
+		ctx.Logger.Errorf("failed to generate authorization info for user: %s, error: %s", ctx.UserID, err)
+		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
+		ctx.UnAuthorized = true
+		return
+	}
 
 	args := new(commonmodels.BasicImage)
 	data, err := c.GetRawData()
@@ -84,6 +117,12 @@ func UpdateBasicImage(c *gin.Context) {
 		log.Errorf("UpdateBasicImage json.Unmarshal err : %v", err)
 	}
 	internalhandler.InsertOperationLog(c, ctx.UserName, "", "更新", "基础镜像", fmt.Sprintf("id:%s", args.ID), string(data), ctx.Logger)
+
+	// authorization checks
+	if !ctx.Resources.IsSystemAdmin {
+		ctx.UnAuthorized = true
+		return
+	}
 
 	c.Request.Body = io.NopCloser(bytes.NewBuffer(data))
 
@@ -97,10 +136,23 @@ func UpdateBasicImage(c *gin.Context) {
 }
 
 func DeleteBasicImage(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
+	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
 
+	if err != nil {
+		ctx.Logger.Errorf("failed to generate authorization info for user: %s, error: %s", ctx.UserID, err)
+		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
+		ctx.UnAuthorized = true
+		return
+	}
+
 	internalhandler.InsertOperationLog(c, ctx.UserName, "", "删除", "基础镜像", fmt.Sprintf("id:%s", c.Param("id")), "", ctx.Logger)
+
+	// authorization checks
+	if !ctx.Resources.IsSystemAdmin {
+		ctx.UnAuthorized = true
+		return
+	}
 
 	ctx.Err = service.DeleteBasicImage(c.Param("id"), ctx.Logger)
 }
